code/requirements.txt
aniso8601==8.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-Reuploaded==0.4.0
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
marshmallow==3.10.0
matplotlib==3.3.3
networkx==2.5
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.0
six==1.15.0
ssim==0.3.0
tifffile==2021.1.11
typing==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
